Add Library for all Modules
