+++ API/pom.xml
@@ -42,6 +42,10 @@
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
     </dependency>
+    <dependency>
+      <groupId>de.brandwatch.minianalytics</groupId>
+      <artifactId>library</artifactId>
+    </dependency>
   </dependencies>
 
   <build>
@@ -59,6 +63,11 @@
             </goals>
           </execution>
         </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>3.0.0-M3</version>
       </plugin>
     </plugins>
   </build>
