+++ API/src/main/java/de/brandwatch/minianalytics/api/controller/MentionController.java
@@ -2,8 +2,8 @@ package de.brandwatch.minianalytics.api.controller;
 
 
 import com.google.gson.Gson;
-import de.brandwatch.minianalytics.api.solr.model.Mention;
-import de.brandwatch.minianalytics.api.solr.repository.MentionRepository;
+import de.brandwatch.minianalytics.library.solr.model.Mention;
+import de.brandwatch.minianalytics.library.solr.repository.MentionRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
