--- API/src/main/java/de/brandwatch/minianalytics/api/solr/config/SolrConfig.java
@@ -1,26 +1,0 @@
-package de.brandwatch.minianalytics.api.solr.config;
-
-import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.impl.HttpSolrClient;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.solr.core.SolrTemplate;
-import org.springframework.data.solr.repository.config.EnableSolrRepositories;
-
-@Configuration
-@EnableSolrRepositories(
-        basePackages = "de.brandwatch.minianalytics.api.solr.repository"
-)
-public class SolrConfig {
-
-    @Bean
-    public SolrClient solrClient(@Value("${spring.data.solr.host}") String solrURL) {
-        return new HttpSolrClient.Builder(solrURL).build();
-    }
-
-    @Bean
-    public SolrTemplate solrTemplate(SolrClient client) {
-        return new SolrTemplate(client);
-    }
-}
