+++ library/src/main/java/de/brandwatch/minianalytics/library/solr/config/SolrConfig.java
@@ -1,26 +1,21 @@
-package de.brandwatch.minianalytics.mentionstorer.solr.config;
+package de.brandwatch.minianalytics.library.solr.config;
 
 import org.apache.solr.client.solrj.SolrClient;
-
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.solr.core.SolrTemplate;
 import org.springframework.data.solr.repository.config.EnableSolrRepositories;
 
 @Configuration
 @EnableSolrRepositories(
-        basePackages = "de.brandwatch.minianalytics.mentionstorer.solr.repository"
-        )
+        basePackages = "de.brandwatch.minianalytics.library.solr.repository"
+)
 public class SolrConfig {
 
-    @Value("${spring.data.solr.host}")
-    String solrURL;
-
     @Bean
-    public SolrClient solrClient() {
+    public SolrClient solrClient(@Value("${spring.data.solr.host}") String solrURL) {
         return new HttpSolrClient.Builder(solrURL).build();
     }
