+++ mentions-generator/src/main/java/de/brandwatch/minianalytics/mentiongenerator/MentionsGeneratorApplication.java
@@ -6,11 +6,13 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
 import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 
@@ -19,6 +21,8 @@ import java.time.Instant;
 @SpringBootApplication
 @EnableCaching
 @EnableScheduling
+@EntityScan(basePackages = {"de.brandwatch.minianalytics.library.postgres.model"})
+@EnableJpaRepositories(basePackages = {"de.brandwatch.minianalytics.library.postgres.repository"})
 public class MentionsGeneratorApplication {
 
     private static final Logger logger = LoggerFactory.getLogger(MentionsGeneratorApplication.class);
