+++ mentions-generator/src/main/java/de/brandwatch/minianalytics/mentiongenerator/kafka/Producer.java
@@ -1,6 +1,6 @@
 package de.brandwatch.minianalytics.mentiongenerator.kafka;
 
-import de.brandwatch.minianalytics.mentiongenerator.model.Mention;
+import de.brandwatch.minianalytics.library.solr.model.Mention;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
