--- mentions-generator/src/main/java/de/brandwatch/minianalytics/mentiongenerator/model/Resource.java
@@ -1,54 +1,0 @@
-package de.brandwatch.minianalytics.mentiongenerator.model;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.InstantDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.InstantSerializer;
-
-import java.time.Instant;
-
-public class Resource {
-
-    private String author;
-    private String text;
-
-    @JsonSerialize(using = InstantSerializer.class)
-    @JsonDeserialize(using = InstantDeserializer.class)
-    private Instant date;
-
-    public Resource() {
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public Instant getDate() {
-        return date;
-    }
-
-    public void setDate(Instant date) {
-        this.date = date;
-    }
-
-    @Override
-    public String toString() {
-        return "{\n" +
-                "\tauthor: " + getAuthor() + "\n" +
-                "\ttext: " + getText() + "\n" +
-                "\tdate: " + getDate().toString() + "\n" +
-                "}";
-    }
-}
