--- mentions-generator/src/main/java/de/brandwatch/minianalytics/mentiongenerator/postgres/model/Query.java
@@ -1,44 +1,0 @@
-package de.brandwatch.minianalytics.mentiongenerator.postgres.model;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "queries")
-public class Query {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "queryID")
-    private long queryID;
-
-    @Column(name = "query")
-    private String query;
-
-    public Query(String query) {
-        this.query = query;
-    }
-
-    public Query() {
-    }
-
-    public long getQueryID() {
-        return queryID;
-    }
-
-    public void setQueryID(long queryID) {
-        this.queryID = queryID;
-    }
-
-    public String getQuery() {
-        return query;
-    }
-
-    public void setQuery(String query) {
-        this.query = query;
-    }
-
-    @Override
-    public String toString() {
-        return query;
-    }
-}
