+++ pom.xml
@@ -13,6 +13,7 @@
         <module>mentions-generator</module>
         <module>API</module>
         <module>twitterpuller</module>
+        <module>library</module>
     </modules>
 
     <properties>
@@ -84,6 +85,16 @@
                 <groupId>com.google.code.gson</groupId>
                 <artifactId>gson</artifactId>
                 <version>2.8.5</version>
+            </dependency>
+            <dependency>
+                <groupId>de.brandwatch.minianalytics</groupId>
+                <artifactId>librry</artifactId>
+                <version>0.0.1-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>de.brandwatch.minianalytics</groupId>
+                <artifactId>library</artifactId>
+                <version>0.0.1-SNAPSHOT</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
