+++ twitterpuller/src/main/java/de/brandwatch/minianalytics/twitterpuller/TwitterPullerApplication.java
@@ -6,11 +6,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
 import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import twitter4j.Twitter;
@@ -23,6 +25,8 @@ import java.time.Instant;
 @SpringBootApplication
 @EnableScheduling
 @EnableCaching
+@EntityScan(basePackages = {"de.brandwatch.minianalytics.library.postgres.model"})
+@EnableJpaRepositories(basePackages = {"de.brandwatch.minianalytics.library.postgres.repository"})
 public class TwitterPullerApplication {
 
     private static final Logger logger = LoggerFactory.getLogger(TwitterPullerApplication.class);
