--- twitterpuller/src/main/java/de/brandwatch/minianalytics/twitterpuller/model/Resource.java
@@ -1,59 +1,0 @@
-package de.brandwatch.minianalytics.twitterpuller.model;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.InstantDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.InstantSerializer;
-
-import java.time.Instant;
-import java.util.Objects;
-
-public class Resource {
-
-    private String author;
-    private String text;
-
-    @JsonSerialize(using = InstantSerializer.class)
-    @JsonDeserialize(using = InstantDeserializer.class)
-    private Instant date;
-
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public Instant getDate() {
-        return date;
-    }
-
-    public void setDate(Instant date) {
-        this.date = date;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Resource)) return false;
-        Resource resource = (Resource) o;
-        return Objects.equals(getAuthor(), resource.getAuthor()) &&
-                Objects.equals(getText(), resource.getText()) &&
-                Objects.equals(getDate(), resource.getDate());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getAuthor(), getText(), getDate());
-    }
-}
